yolo-segmentation-lab/requirements.txt
ultralytics>=8.3.0
opencv-python
numpy
Pillow
pyyaml
torch
torchvision
torchaudio
rembg
onnxruntime
